Scripts/requirements.txt
dotenv
openai
sklearn.model_selection
json
pandas
requests
matplotlib.pyplot
os